+++ CHANGELOG.md
@@ -92,6 +92,8 @@ TODO
   [#190](https://github.com/scionproto-contrib/jpan/pull/190)
 - Channel `send()` cleanup.
   [#192](https://github.com/scionproto-contrib/jpan/pull/192)
+- Bootstrap with reverse lookup now also considers subnet addresses, support for SOA and SRV..
+  [#194](https://github.com/scionproto-contrib/jpan/pull/194)
 
 ### Fixed
 - Post 0.5.0 cleanup